Working on frame handling. In terms of the QT code, the non-frame-related specs pass with these changes (e.g. mainFrame() => currentFrame() and the new injectJavascriptHelpers() code in WebPage.cpp).
It seems like the current JS+xpath implementation dives in to iframes already.
Is this desired behavior? I wonder if that works with x-domain iframes? I
doubt it...

Also, this design assumes that we only step one frame down at a time...

Lastly, I'm really not sure how QWebKit decides which frame is currentFrame().
For now, I'm hoping to be able to use the QWebFrame setFocus() method. This
may be a dead end though. We may have to have WebPage manually keep track of
the "current" frame.