Post screenshots to S3.
When screenshots are taken in CI environments, the files are saved to
temporary directories and are not accessible. This commit allows the
posting of screenshots to an S3 bucket so that they can be accessed
after CI builds.

Posting to S3 will happen automatically for screenshots if the
environment variable `CAPYBARA_WEBKIT_S3_SCREENSHOTS_BUCKET` is set.