Update spec to condition that fails on master
Needed to rebuild the capybara binary to properly get this spec to fail when
testing on master. When running with the patched binary (the one built from this
branch) the previous version of this spec would fail, but it would actually
pass against master when running with master's version of the binary. The spec
now correctly reflects the failure condition (which is a colon followed by a
space ": "); it fails when run on master with master's capybara binary and
passes on this branch with the updated binary.