Make caching behavior more like Selenium
We disabled caching entirely, because it sometimes led to unexpected
behavior in tests, and Qt 4.8 didn't behave the same as Qt 5.x.

However, this introduced issues:

* Selenium does perform caching, so capybara-webkit behaved differently
  when switching drivers
* Without any caching, assets were sometimes requested twice for the
  same page

This commit re-enables caching and attempts to improve the way we behave
on cached pages so that testing is less unpredictable. It adds tests to
ensure we continue to behave similarly to Selenium.