Remove auto-screenshot behavior
We can't distinguish between ClickFailed errors which will be retried,
and those which have actually failed for good. This means we end up
writing hundreds of screenshot files during a failure, and we sometimes
write several screenshot files even when the click ends up working.

This isn't feasible to work around in a driver, so this commit removes
the behavior from capybara-webkit. Something like this may make more
sense in Capybara proper.

Resolves #626.