Hide CDPATH from make
The CDPATH environment variable, as used by zsh and GNU Bash, interacts
poorly with make, so hide it.

To see it interact poorly, prior to this diff:

    mkdir /tmp/src /tmp/test
    export CDPATH=/tmp
    rake build

When it attempts to `cd src`, it will go to `/tmp/src` instead of the
correct place in this tree.

Examples of this bug have appeared in #56, #572, and #852.

Closes #947.